doso-ai/requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.14.1
sqlalchemy>=2.0.15
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
openai>=1.3.0
scikit-learn>=1.2.2
psycopg2-binary>=2.9.6
redis>=4.5.5
prophet>=1.1.4
pytest>=7.3.1
